backend/requirements.txt
# Django and core dependencies
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-extensions>=3.2.0

# Authentication
djangorestframework-simplejwt>=5.2.0

# Database
psycopg2-binary>=2.9.0
mysqlclient>=2.1.0

# File handling and media
Pillow>=10.0.0

# PDF Generation
reportlab>=4.0.0

# Excel handling
openpyxl>=3.1.0

# Celery for background tasks
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
flower>=2.0.0

# WebSocket support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# History tracking
django-simple-history>=3.4.0
django-auditlog>=2.0.0

# Email
django-anymail>=10.0.0

# Development
python-decouple>=3.8.0

# Other utilities
python-dateutil>=2.8.0
geoip2>=4.6.0
user-agents>=2.2.0
